[+] add : share modal component
